# web2/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Clubs API",
    "description": "A simple API for managing clubs",
    "contact": {
      "name": "Karlo Franić",
      "email": "karlofranic0@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "http://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
    "/docs": {
      "get": {
        "summary": "Retrieve a list of all clubs",
        "responses": {
          "200": {
            "description": "Json representation of this document"
          }
        }
      }
    },
    "/clubs": {
      "get": {
        "summary": "Retrieve a list of all clubs",
        "responses": {
          "200": {
            "description": "A list of clubs",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Club"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new club",
        "requestBody": {
          "content": {
            "*/*": {
              "schema": {
                "$ref": "#/components/schemas/Club"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The created club",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Club"
                }
              }
            }
          },
          "500": {
            "description": "Error creating club",
            "content": {}
          }
        },
        "x-codegen-request-body-name": "body"
      }
    },
    "/clubs/{id}": {
      "get": {
        "summary": "Retrieve a club by its ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The ID of the club to retrieve",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A single club",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Club"
                }
              }
            }
          },
          "404": {
            "description": "Club not found",
            "content": {}
          }
        }
      },
      "put": {
        "summary": "Update a club",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The ID of the club to update",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "*/*": {
              "schema": {
                "$ref": "#/components/schemas/Club"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The updated club",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Club"
                }
              }
            }
          },
          "404": {
            "description": "Club not found",
            "content": {}
          },
          "500": {
            "description": "Error updating club",
            "content": {}
          }
        },
        "x-codegen-request-body-name": "body"
      },
      "delete": {
        "summary": "Delete a club",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The ID of the club to delete",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The deleted club",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Club"
                }
              }
            }
          },
          "404": {
            "description": "Club not found",
            "content": {}
          },
          "500": {
            "description": "Error deleting club",
            "content": {}
          }
        }
      }
    },
    "/clubs/city/{city}": {
      "get": {
        "summary": "Retrieve a list of clubs in a specific city",
        "parameters": [
          {
            "name": "city",
            "in": "path",
            "description": "The city of the clubs to retrieve",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of clubs",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Club"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No clubs found in the specified city",
            "content": {}
          }
        }
      }
    },
    "/clubs/name/{name}": {
      "get": {
        "summary": "Search for clubs by name",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "The name of the clubs to search for",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of clubs",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Club"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No clubs found with the specified name",
            "content": {}
          }
        }
      }
    },
    "/clubs/rating/{rating}": {
      "get": {
        "summary": "Retrieve a list of clubs with a minimum rating",
        "parameters": [
          {
            "name": "rating",
            "in": "path",
            "description": "The minimum rating of the clubs to retrieve",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of clubs",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Club"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/clubs/top/{top}": {
      "get": {
        "summary": "Retrieve the top N clubs by rating",
        "parameters": [
          {
            "name": "top",
            "in": "path",
            "description": "The number of top clubs to retrieve",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of clubs",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Club"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Club": {
        "type": "object",
        "properties": {
          "club_id": {
            "type": "integer"
          },
          "club_name": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "rating": {
            "type": "number"
          },
          "address": {
            "type": "string"
          },
          "phone_number": {
            "type": "string"
          },
          "instagram": {
            "type": "string"
          },
          "work_hours": {
            "type": "integer"
          },
          "music_genre": {
            "type": "string"
          },
          "bottle_service": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      }
    }
  }
}